Unit Test For Database File
This unit test does a simple but very important feature check, which would be whether we are actually calling out and connecting to a database file. Specifically the “Liquor_Database.db” we are meant to be using. It initially checks whether the name of the file exists inside the path we give it. The second requirement to make this whole test pass is to check whether the file is a database type of file. Each requirement has its own exception handling as needed as well. There is a SQL query that is run that specifically takes a look at the very first table available regardless of the database using a call that calls onto “sqlite_master” which should be able to correctly reference if it's executed onto a database type file. As of now the code is correctly working and passes all the tests.